+++ R/data.R
@@ -42,3 +42,23 @@
 #'
 #' @source <https://github.com/sysbio-bioinf/VennMaster/blob/master/data_examples/deploy/example1.list>
 "organisms"
+
+#' Pain distribution data
+#'
+#' Data from a study on pain distribution for patients with
+#' persistent neck pain in relation to a whiplash trauma.
+#'
+#' @section Disclaimer:
+#' Note that the maintainer of this package is an author of the source for
+#' this data.
+#'
+#' @format A flat [table] (cross-table) with with sex in columns and
+#'   pain distribution in rows and integer counts making up the cells of the
+#'   table.
+#'
+#' @source Westergren H, Larsson J, Freeman M, Carlsson A, Jöud A,
+#'   Malmström E-M. Sex-based differences in pain distribution in a cohort of
+#'    patients with persistent post-traumatic neck pain. Disability and
+#'    Rehabilitation [Internet]. 2017 Jan 27 [cited 2017 Jan 28]; Available
+#'    from: <http://www.tandfonline.com/eprint/RK8tusJswJCsa9YakGDW/full>
+"pain"